+++ src/main/java/be/elmital/highlightItem/mixin/AbstractContainerScreenMixin.java
@@ -26,17 +26,24 @@ import be.elmital.highlightItem.Colors;
 import be.elmital.highlightItem.Configurator;
 import be.elmital.highlightItem.HighlightItem;
 import be.elmital.highlightItem.ItemComparator;
+import be.elmital.highlightItem.UnsupportedMinecraftClassOperationException;
+import be.elmital.highlightItem.Utils;
 import com.llamalad7.mixinextras.sugar.Local;
 import net.fabricmc.api.EnvType;
 import net.fabricmc.api.Environment;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphicsExtractor;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
+import net.minecraft.client.gui.screens.inventory.CreativeModeInventoryScreen;
 import net.minecraft.client.input.KeyEvent;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.TextColor;
+import net.minecraft.world.inventory.AbstractContainerMenu;
 import net.minecraft.world.inventory.Slot;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyArgs;
@@ -52,6 +59,29 @@ public abstract class AbstractContainerScreenMixin {
 
 	@Shadow @Nullable protected Slot hoveredSlot;
 
+	@Unique private boolean highlightItemCompatible;
+
+	@Inject(method = "<init>(Lnet/minecraft/world/inventory/AbstractContainerMenu;Lnet/minecraft/world/entity/player/Inventory;Lnet/minecraft/network/chat/Component;II)V", at	= @At(value = "TAIL"))
+	private void construct(CallbackInfo ci, @Local(argsOnly = true) AbstractContainerMenu menu) {
+		// Check possible incompatibility
+		try {
+			Utils.isInPlayerInventory(menu.getSlot(0), (AbstractContainerScreen<? extends AbstractContainerMenu>) (Object) this);
+			this.highlightItemCompatible = true;
+		} catch (UnsupportedMinecraftClassOperationException e) {
+			this.highlightItemCompatible = false;
+			HighlightItem.LOGGER.error(e);
+		} catch (UnsupportedOperationException e) {
+			// Change Screen context value if the Screen context is set to non default value
+			if (!Configurator.SCREEN_CONTEXT.equals(Configurator.ScreenContext.EVERYWHERE) || !Configurator.SCREEN_CONTEXT.equals(Configurator.ScreenContext.EXCLUDE_CREATIVE)) {
+				Configurator.SCREEN_CONTEXT = Configurator.ScreenContext.EXCLUDE_CREATIVE;
+				Minecraft.getInstance().player.sendSystemMessage(Component.literal("The option for Screen limitation have been deactivated due to a compatibility issue! Please, check your logs and report it to the HighLightItem issue tracker.").withColor(TextColor.RED));
+			}
+
+			this.highlightItemCompatible = false;
+			HighlightItem.LOGGER.error(e);
+		}
+	}
+
 	@Inject(method = "extractSlots", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screens/inventory/AbstractContainerScreen;extractSlot(Lnet/minecraft/client/gui/GuiGraphicsExtractor;Lnet/minecraft/world/inventory/Slot;II)V", shift = At.Shift.AFTER))
 	private void drawSlot(GuiGraphicsExtractor guiGraphics, int i, int j, CallbackInfo ci, @Local Slot slot) {
 		if (Configurator.TOGGLE) {
@@ -61,12 +91,41 @@ public abstract class AbstractContainerScreenMixin {
 			if (hoveredSlot.equals(slot) && !Configurator.COLOR_HOVERED)
 				return;
 
-			if (slot.isActive() && !slot.getItem().isEmpty() && ItemComparator.test(Configurator.COMPARATOR, hoveredSlot.getItem(), slot.getItem())) {
+			if (!slot.isActive() || slot.getItem().isEmpty())
+				return;
+
+			if (shouldSkip(slot))
+				return;
+
+			if (ItemComparator.test(Configurator.COMPARATOR, hoveredSlot.getItem(), slot.getItem())) {
 				HighlightItem.toDrawFromMod = slot;
 				extractSlotHighlightFront(guiGraphics);
 				HighlightItem.toDrawFromMod = null;
 			}
 		}
+	}
+
+	@Unique
+    @SuppressWarnings("ConstantConditions")
+	private boolean shouldSkip(Slot slot) {
+		if (!this.highlightItemCompatible)
+			return false;
+
+		if (Configurator.SCREEN_CONTEXT.equals(Configurator.ScreenContext.EVERYWHERE))
+			return false;
+
+		if (Configurator.SCREEN_CONTEXT.excludeCreativeScreen() && CreativeModeInventoryScreen.class.isInstance(this))
+			return true;
+
+		try {
+			if (Configurator.SCREEN_CONTEXT.inContainer() && Utils.isInPlayerInventory(slot, (AbstractContainerScreen<? extends AbstractContainerMenu>) (Object) this))
+				return true;
+
+			if (Configurator.SCREEN_CONTEXT.inPlayerInventoryPart() && !Utils.isInPlayerInventory(slot, (AbstractContainerScreen<? extends AbstractContainerMenu>) (Object) this))
+				return true;
+		} catch (UnsupportedOperationException _) {}
+
+		return false;
 	}
 
 	@ModifyArgs(method = "extractSlotHighlightFront", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/GuiGraphicsExtractor;blitSprite(Lcom/mojang/blaze3d/pipeline/RenderPipeline;Lnet/minecraft/resources/Identifier;IIII)V"))